requirements.txt
aiogram-dialog==2.0.0
pytest==7.4.2
PyYAML==6.0.1
pyyaml-include==1.3.1
pydantic==2.3.0